test GITU

+++ resources/views/cisteni/index.blade.php
@@ -8,7 +8,7 @@
                 {{ $jednoCisteni->id }} - {{ $jednoCisteni->idVybehu }} výběh v {{ $jednoCisteni->casCisteni }}
             </a>
             @level(2)
-            {{ Form::open(['method' => 'DELETE', 'route' => ['cisteni.destroy', $jednoCisteni->id]]) }}
+            {{ Form::open(['method' => 'DELETE', 'route' => ['cisteni.destroy', $jednoCisteni->id], 'onsubmit' => 'return ConfirmDelete()']) }}
             {{ Form::submit('Delete', ['class' => 'btn btn-danger']) }}
             {{ Form::close() }}
             @endlevel
