+++ resources/views/osetrovatele/index.blade.php
@@ -8,7 +8,7 @@
                 {{ $osetrovatel->jmeno . ' ' . $osetrovatel->prijmeni }}
             </a>
             @level(2)
-            {{ Form::open(['method' => 'DELETE', 'route' => ['osetrovatele.destroy', $osetrovatel->id]]) }}
+            {{ Form::open(['method' => 'DELETE', 'route' => ['osetrovatele.destroy', $osetrovatel->id], 'onsubmit' => 'return ConfirmDelete()']) }}
             {{ Form::submit('Delete', ['class' => 'btn btn-danger']) }}
             {{ Form::close() }}
             @endlevel
@@ -20,11 +20,11 @@
     {{--TODO @iis tohle tu musi byt kvůli tlačítku - nestaci to dát do layoutu - proč?--}}
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
 
-    {{--@level(2)--}}
-        {{--<a href="/osetrovatele/create">--}}
-            {{--<span class="glyphicon glyphicon-plus"></span>--}}
-        {{--</a>--}}
-    {{--@endlevel--}}
+    @level(3)
+        <a href="{{url('/osetrovatele/create')}}">
+            <span class="glyphicon glyphicon-plus"></span>
+        </a>
+    @endlevel
