+++ resources/views/vybehy/index.blade.php
@@ -5,10 +5,14 @@
     @foreach($vybehy as $vybeh)
         <li>
             <a href="{{url('/vybehy'). '/' . $vybeh->id }}">
-                {{ $vybeh->id }}
+                @foreach($typyVybehu as $typ)
+                    @if($vybeh->idTypuVybehu === $typ->id)
+                        {{ $vybeh->id }} - {{$typ->nazev}}
+                    @endif
+                @endforeach
             </a>
             @level(3)
-            {{ Form::open(['method' => 'DELETE', 'route' => ['vybehy.destroy', $vybeh->id]]) }}
+            {{ Form::open(['method' => 'DELETE', 'route' => ['vybehy.destroy', $vybeh->id], 'onsubmit' => 'return ConfirmDelete()']) }}
             {{ Form::submit('Delete', ['class' => 'btn btn-danger']) }}
             {{ Form::close() }}
             @endlevel
